// src/screens/tools/effects/CSSFilters/types.ts
export type TFilterConfig = {
  blur: number;
  brightness: number;
  contrast: number;
  grayscale: number;
  hueRotate: number;
  invert: number;
  saturate: number;
  sepia: number;
  opacity: number;
  dropShadowX: number;
  dropShadowY: number;
  dropShadowBlur: number;
  dropShadowColor: string;
};

export type TPreviewImage = {
  name: string;
  url: string;
};